Refactor module and node implementation
