+++ native/src/core/core.hpp
@@ -14,7 +14,7 @@ std::string read_certificate(int fd, int version = -1);
 
 // Module stuffs
 void handle_modules();
-void magic_mount();
+void load_modules();
 void disable_modules();
 void remove_modules();
 void exec_module_scripts(const char *stage);
